Rename InputTutorial directory and files
